requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Markdown==3.1.1
django-filter==2.0.0
djangorestframework==3.9.4
django-ckeditor==5.9.0
Pillow==7.1.2
django-cors-headers==2.4.0
django-seed==0.2.2